// app/model/Task.tsx
export type Task = {
  id: number;
  title: string;
  dueDate: number;
  category: string;

  // constructor(id: string, title: string, dueDate: number, category: string) {
  //     this.id = id;
  //     this.title = title;
  //     this.dueDate = dueDate;
  //     this.category = category;
  // }
};